+++ juniper_graphql_ws/src/graphql_transport_ws/mod.rs
@@ -39,7 +39,7 @@ pub use self::{
 
 struct ExecutionParams<S: Schema> {
     subscribe_payload: SubscribePayload<S::ScalarValue>,
-    config: Arc<ConnectionConfig<S::Context>>,
+    config: ConnectionConfig<S::Context>,
     schema: S,
 }
 
@@ -81,7 +81,7 @@ enum ConnectionState<S: Schema, I: Init<S::ScalarValue, S::Context>> {
     PreInit { init: I, schema: S },
     /// Active is the state after a ConnectionInit message has been accepted.
     Active {
-        config: Arc<ConnectionConfig<S::Context>>,
+        config: ConnectionConfig<S::Context>,
         stoppers: HashMap<String, oneshot::Sender<()>>,
         schema: S,
     },
@@ -89,7 +89,12 @@ enum ConnectionState<S: Schema, I: Init<S::ScalarValue, S::Context>> {
     Terminated,
 }
 
-impl<S: Schema, I: Init<S::ScalarValue, S::Context>> ConnectionState<S, I> {
+impl<S, I> ConnectionState<S, I>
+where
+    S: Schema,
+    S::Context: Clone,
+    I: Init<S::ScalarValue, S::Context>,
+{
     // Each message we receive results in a stream of zero or more reactions. For example, a
     // Ping message results in a one-item stream with the Pong message reaction.
     async fn handle_message(
@@ -121,7 +126,7 @@ impl<S: Schema, I: Init<S::ScalarValue, S::Context>> ConnectionState<S, I> {
 
                         (
                             Self::Active {
-                                config: Arc::new(config),
+                                config,
                                 stoppers: HashMap::new(),
                                 schema,
                             },
@@ -489,6 +494,7 @@ where
     T: TryInto<Input<S::ScalarValue>>,
     T::Error: Error,
     S: Schema,
+    S::Context: Clone,
     I: Init<S::ScalarValue, S::Context> + Send,
 {
     type Error = Infallible;