+++ juniper_graphql_ws/CHANGELOG.md
@@ -6,6 +6,18 @@ All user visible changes to `juniper_graphql_ws` crate will be documented in thi
 
 
 
+## [0.6.0] · 2026-03-??
+[0.6.0]: /../../tree/juniper_graphql_ws-v0.6.0/juniper_graphql_ws
+
+### BC Breaks
+
+- Context now required `Clone` bound. ([#1369])
+
+[#1369]: /../../pull/1369
+
+
+
+
 ## [0.5.0] · 2025-09-08
 [0.5.0]: /../../tree/juniper_graphql_ws-v0.5.0/juniper_graphql_ws